config play load & update token on call
