+++ project/backend/app/appuac/service/auth.py
@@ -12,6 +12,7 @@ def token_auth(token: str):
     if auth_session.is_expired:
         auth_session.delete()
         return None
+    auth_session.refresh()
     auth_session.user.last_login = auth_session.last_used
     auth_session.user.save()
     return auth_session
@@ -22,5 +23,4 @@ class BearerTokenAuth(HttpBearer):
     Bearer TOken Cass
     """
     def authenticate(self, request: HttpRequest, token: str) -> Any | None:
-        
         return token_auth(token)
